# Compiler args performance enhancements