See #1485 for test case; I assume there are no performance test cases at the moment. Without running the profiler, speeds are little bit better than I said before (averaged over 10 runs):
* master 80f870c: 18.05±0.05s
* ~~Use an `OrderedDict` when deduping d07b22f: 3.52±0.05s (×5.12±0.07)~~
* Use a `list`+`set` 128e2a6: 3.52±0.07s (×5.12±0.10)
* With a cache for basic compiler args ~~6d9d268~~ 11f8d4a: 2.52±0.04s (×7.16±0.13)

Profile now looks like this:
![1000call-perf](https://cloud.githubusercontent.com/assets/302469/24129661/821d2d46-0dba-11e7-8c38-57d09fd5dd6d.png)

You can make a _very_ small improvement (2.52±0.03s (×7.17±0.08)) by replacing `arg.startswith(prefixes)` with `arg[:2] in prefixes` in the `_can_dedup`/`_should_prepend` functions, but this assumes that all prefixes are 2 characters.

The caching is only based on global/subproject information; if the cache were made from all target properties, I'm sure things could be even faster, but I didn't feel like puzzling out what of the target to use in the key (since obviously using the target itself would be useless.)