@QuLogic 2017-03-21
Err, why does the test case require that duplicate args be duplicated?