@QuLogic 2017-03-21
Sigh, I just realized the dict is over-dedup'ing. I might have to go with a `set`+`list` depending on how it times out.