@QuLogic 2017-03-21
`list`+`set` seems pretty close: 2.63±0.13s