@codido 2017-03-21
@QuLogic Thanks!

With the original example I'm seeing a huge boost, though as you mentioned, it still takes a couple of seconds. Not sure what would be a reasonable/feasible time, but running the same example with a single C file takes less than half a second, so guess things aren't completely cached just yet?

Also, for multiple include directories, such as the following example, I'm afraid I'm not seeing much difference with this (actually even runs a bit slower):
```
#!/bin/bash

cat > meson.build <<EOF
project('example', 'c')
inc = []
src = []
EOF

for i in {1..1000}
do
   mkdir -p dir$i
   touch file$i.c
   echo "inc += include_directories('dir$i')" >> meson.build
   echo "src += files('file$i.c')" >> meson.build
done

echo "executable('example', src, include_directories: inc)" >> meson.build

time meson build
rm -rf file*.c dir* build
```