@QuLogic 2017-03-21
Yes, it only caches based on compiler and subproject, but now that I think about it, maybe it can cache based on target as well if it needs to be applied to all the sources of the target.