@QuLogic 2017-03-21
Ah, I see it's triggered by extraneous stuff at a higher level; adding a cache at that point also helps: ~2.3s. I don't have time to totally test it now; maybe that cache is enough and the other changes aren't necessary...

Most of the time now is in output, which is full of string replacement for escaping. Maybe with `File` objects everywhere, this could be optimized away, but that's a large project.