@nirbheek 2017-03-21
In addition to creating all these caches, I wonder if it makes sense to create per-target rules in the `build.ninja` file. Those are for exactly this purpose, and it would also:

1. Speed up writing of the Ninja file
2. Reduce the size of the `build.ninja` file enormously
3. Improve Ninja's parsing of the `build.ninja` file

We could in fact make things even nicer by having a rule like:

```ninja
incpaths = -Ifoo -Ibar
libpaths = -Lfdir -Lbdir
args = -e -v -e -r -y -t -h -i -n -g -else
rule sometarget
  command = gcc $incpaths $libpaths $in $args -o $out 
```

or similar.

(moved from the issue to here)