@QuLogic 2017-03-23
I had a chance to try out just the last commit, and since it's a top-level cache, both this case and the other issue are about the same, around 2.2 seconds. It's about the same or slightly slower (a few hundredths of a second) when the first two commits are also included. I didn't look at a spread of multiple runs to get any standard deviation or anything. Odd that it seems a little faster than last time, but I think this just getting into external noise at this point.

So I'm not sure whether we need to hang on to the first two commits, or whether they might be of some benefit for other pathological cases.