@jpakkane 2017-03-23
Simpler is better. If we can't get measurable proof that the changes make things faster, let's just drop them.

Per-target rules might make sense, but they complicate creating the compilation database a bit as it is generated for rules. But let's do that in a different MR in any case.