@QuLogic 2017-03-23
I'll hang on to them in case we need them, but I've removed the others from this PR.